# optimized brifand::fold on gcc