Tips: Types members impacts the performance of gcc.

----

```cpp
using list = brigand::fold<
  brigand::range<int, 0, N_L1>,
  std::integral_constant<int, 0>,
  brigand::plus<brigand::_state, brigand::_element>
>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 37M | 00.10s - 39M | 00.15s - 42M | 00.24s - 49M | 00.44s - 63M
gcc | 00.05s - 21M | 00.10s - 27M | 00.16s - 35M | 00.28s - 53M | 00.55s - 93M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.07s - 36M | 00.10s - 39M | 00.15s - 42M | 00.24s - 49M | 00.44s - 63M
gcc | 00.05s - 21M | 00.10s - 27M | 00.15s - 34M | 00.26s - 50M | 00.50s - 84M